requirements.txt
easyocr==1.7.0
Faker==19.1.0
jaconv==0.3.4
largestinteriorrectangle==0.2.0
mss==9.0.1
numpy==1.24.3
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
Pillow==9.5.0
Pillow==10.0.1
pycountry==22.3.5
PyHyphen==4.0.3
PySimpleGUI==4.60.5
pytesseract==0.3.10
Requests==2.31.0
setuptools==67.8.0
simple_lama_inpainting==0.1.0
timm==0.9.2
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tornado==6.3.2
tqdm==4.65.0
ultralytics==8.0.118
